refactor(command): 重构命令处理系统
- 添加了命令冷却时间管理功能，支持全局、机器人、群组和用户级别的冷却
- 重构了命令匹配器，支持别名和前缀匹配，提升命令识别准确性
- 添加了好友和群组配置管理，支持独立的冷却时间和反应模式设置
- 优化了BotContext和MessageContext的实现，使用Arc共享减少内存拷贝
- 移除了旧的Matcher trait，统一使用Handler trait进行事件处理
- 更新了连接时间处理逻辑，使用DateTime替代SystemTime提高时间处理精度
- 重新组织了配置模块结构，添加了ReactiveMode枚举和FriendConfig支持
- 集成了chrono库用于更精确的时间戳处理和冷却时间计算

+++ crates/puniyu_config/src/group/option.rs
@@ -1,4 +1,4 @@
-use crate::bot::option::ReactiveMode;
+use crate::common::ReactiveMode;
 use serde::{Deserialize, Serialize};
 
 /// TODO: 群组中的禁用/启用插件
@@ -6,10 +6,10 @@ use serde::{Deserialize, Serialize};
 pub struct GroupOption {
 	/// 群组cd冷却时间
 	#[serde(default = "default_group_cd")]
-	cd: u8,
+	cd: u64,
 	/// 用户cd冷却时间
 	#[serde(default = "default_group_user_cd")]
-	user_cd: u8,
+	user_cd: u64,
 	#[serde(default = "default_reactive_mode")]
 	/// 响应模式
 	mode: ReactiveMode,
@@ -30,11 +30,11 @@ impl Default for GroupOption {
 	}
 }
 
-fn default_group_cd() -> u8 {
+fn default_group_cd() -> u64 {
 	0
 }
 
-fn default_group_user_cd() -> u8 {
+fn default_group_user_cd() -> u64 {
 	0
 }
 
@@ -48,11 +48,11 @@ fn default_alias() -> Vec<String> {
 
 impl GroupOption {
 	/// 获取群组cd冷却时间
-	pub fn cd(&self) -> u8 {
+	pub fn cd(&self) -> u64 {
 		self.cd
 	}
 	/// 获取用户cd冷却时间
-	pub fn user_cd(&self) -> u8 {
+	pub fn user_cd(&self) -> u64 {
 		self.user_cd
 	}
 	/// 获取响应模式
